+++ src/test/java/com/unimag/Tienda/Unit/RepositoryTest/PedidoRepositoryTest.java
@@ -1,0 +1,4 @@
+package com.unimag.Tienda.Unit.RepositoryTest;
+
+public class PedidoRepositoryTest {
+}
