--- src/test/java/com/unimag/Tienda/repositoryTest/ItemPedidoRepositoryTest.java
@@ -1,4 +1,0 @@
-package com.unimag.Tienda.repositoryTest;
-
-public class ItemPedidoRepositoryTest {
-}
