--- src/test/java/com/unimag/Tienda/repositoryTest/PedidoRepositoryTest.java
@@ -1,4 +1,0 @@
-package com.unimag.Tienda.repositoryTest;
-
-public class PedidoRepositoryTest {
-}
